move webview into flare-handler for build
